Split files into managable sizes
