+++ sc/tw/src/main/scala/threaded/ConfigClasses.scala
@@ -1,0 +1,19 @@
+package org.lamastex.mep.tw
+
+case class StreamConfig (
+  handlesFilePath: String,
+  streamDuration: Long
+)
+
+case class WriteConfig (
+  outputFilenames: String,
+  fullFilesDirectory: String,
+  maxFileSize: Long,
+  writeRate: Long
+)
+
+case class TweetSchema (
+  id: Long, 
+  json: String, 
+  time: Long
+)
